requirements.txt
accelerate==0.15.0
datasets==2.7.1
evaluate==0.3.0
faiss_gpu==1.7.2
nltk==3.8
numpy==1.23.4
openai==0.27.1
rank_bm25==0.2.2
requests==2.28.1
scikit_learn==1.2.1
sentence_transformers==2.2.2
torch>=1.13.1
tqdm==4.64.1
transformers